+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-# 0.1.4
+# 1.0.0
+
+Supported the redirect v/<video id>
+
+Added videohub player css
+
+Stopped player from opening new window on endcard links
+
+# 0.1.5
 
 Included the latest videojs-ga that allows for prefixing with
  tracker names.
@@ -9,6 +17,7 @@ Included the latest videojs-ga that allows for prefixing with
 new VidePlayer(videoEl, { pluginConfig: { ga: { gaPrefix: 'trackerpreifx' } });
 ```
 
+# 0.1.4
 # 0.1.3
 
 * BAD BUILD *